apps/users/filters.py
import django_filters

from apps.users.models import User


class UserFilter(django_filters.FilterSet):
    first_name = django_filters.CharFilter(lookup_expr='icontains')

    class Meta:
        model = User
        fields = []

