requirements/requirements.txt
absl-py==0.9.0
amqp==2.3.2
appdirs==1.4.3
arrow==0.12.1
astor==0.7.1
astroid==2.4.1
astunparse==1.6.3
backcall==0.1.0
beautifulsoup4==4.6.3
billiard==3.5.0.4
boto==2.49.0
boto3==1.7.84
botocore==1.10.84
cachetools==4.1.0
celery==4.2.1
certifi==2018.8.13
chardet==3.0.4
cloudpickle==0.6.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cycler==0.10.0
Cython==0.29.13
dask==0.20.0
decorator==4.3.0
defusedxml==0.5.0
distro==1.3.0
Django==2.1
django-appconf==1.0.2
django-authtools==1.6.0
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-filter==2.0.0
django-imagekit==4.0.2
django-model-utils==3.1.2
django-rest-swagger==2.1.0
django-ses==0.8.6
django-six==1.0.4
djangorestframework==3.8.2
djangorestframework-bulk==0.2.1
djangorestframework-filters==0.10.2
djangorestframework-xml==1.3.0
docutils==0.14
et-xmlfile==1.0.1
Faker==0.9.0
future==0.16.0
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
html5lib==1.0.1
idna==2.7
imagecodecs==2020.2.18
imageio==2.8.0
inflection==0.4.0
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.10
jmespath==0.9.3
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kombu==4.2.1
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mysqlclient==1.4.6
networkx==2.4
numpy==1.18.4
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.3
parso==0.7.0
pdf2image==1.13.1
pdfminer.six==20200517
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
pilkit==2.0
Pillow==7.1.2
prompt-toolkit==3.0.5
protobuf==3.12.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.4
python-dateutil==2.7.3
python-docx==0.8.10
python-stdnum==1.13
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
redis==3.5.2
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.1.13
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.4.1
selectivesearch==0.4
simplejson==3.17.0
six==1.14.0
sklearn==0.0
sortedcontainers==2.1.0
suds-py3==1.4.1.0
tabula-py==2.1.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0
termcolor==1.1.0
text-unidecode==1.2
threadpoolctl==2.0.0
tifffile==2020.5.11
toml==0.10.1
toolz==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
Wand==0.6.1
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xhtml2pdf==0.2.4
xlrd==1.2.0
XlsxWriter==1.2.8
xmltodict==0.12.0
zxcvbn==4.4.28
zxcvbn-python==4.4.24
